BigDecimal serialization

+++ src/main/java/com/foodit/test/sample/controller/message/SalesAmount.java
@@ -1,14 +1,16 @@
 package com.foodit.test.sample.controller.message;
 
+import java.math.BigDecimal;
+
 public class SalesAmount {
 
-	private final double salesAmount;
+	private final BigDecimal salesAmount;
 
-	public SalesAmount(double salesAmount) {
+	public SalesAmount(BigDecimal salesAmount) {
 		this.salesAmount = salesAmount;
 	}
 
-	public double getSalesAmount() {
+	public BigDecimal getSalesAmount() {
 		return this.salesAmount;
 	}
